#166 Refactor: 푸터 통일
